requirements.txt
# Data Collection and Processing
beautifulsoup4==4.12.2
requests==2.31.0
pandas>=2.2.0
pytesseract==0.3.10
pdf2image==1.16.3
openpyxl==3.1.2

# Data Analysis and Machine Learning
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost==2.0.0
prophet==1.1.4
statsmodels>=0.14.1

# Network Analysis
networkx==3.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Natural Language Processing
spacy>=3.7.2
textblob==0.17.1
transformers==4.35.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
gensim==4.3.0

# Visualization
matplotlib==3.7.2
plotly==5.18.0
folium==0.14.0
seaborn==0.12.2

# API and Web
flask==2.3.3
gunicorn==21.2.0

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
redis==5.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
responses==0.23.0
faker==19.6.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
pytz==2023.3
pyyaml==6.0.0
tenacity==8.2.0

# Monitoring
prometheus-client==0.17.0
sentry-sdk==1.31.0

# PDF Generation
reportlab==4.0.4
pdfkit==1.0.0
weasyprint==60.1

# Documentation
sphinx==7.2.0
sphinx-rtd-theme==1.3.0
